Module jormungandr_lib::interfaces::config::node

Structs

Bootstrap

Bootstrap contains meta data for initial startup

Connection

Miscellaneous network configuration

Cors

CorsOrigin

CorsOriginVisitor 🔒

HeaderName

HeaderNameVisitor 🔒

HttpMethod

JRpc

LayersConfig

Jörmungandr provides multiple additional layers to the poldercast default ones: the preferred list or the bottle in the sea.

NodeConfig

P2p

Policy

policy module

PreferredListConfig

PreferredViewMax

Rest

Tls

TopicsOfInterest

TrustedPeer

Configuration item for a trusted peer.

Visitor 🔒

Constants

DEFAULT_PREFERRED_VIEW_MAX 🔒

Type Definitions

NodeId

Identifier of a peer node.