Struct jormungandr_lib::interfaces::Certificate

pub struct Certificate(pub Certificate);

Tuple Fields
0: Certificate
Implementations

impl Certificate

Use bech32m variant to serialize a certificate as its length it’s not fixed but allow to read original bech32 formatted certificates for backward compatibility

pub fn to_bech32m(&self) -> Result<String, CertificateToBech32Error>

pub fn from_bech32(bech32: &str) -> Result<Self, CertificateFromBech32Error>